Full Search Bar functionality on mian cards page
