+++ client/src/components/SearchBar/index.jsx
@@ -4,22 +4,22 @@ import { useState } from "react";
 const Search = () => {
     const [results, setResults] = useState([]);
   
-    const handleSearch = async (query) => {
-      console.log(query);
-      if (query.trim() === '') return;
-      try {
-        const response = await fetch(`https://api.lorcana-api.com/cards/fetch?search=Name~${query}`);
-        const data = await response.json();
-        console.log(data);
-        setResults(data);
-      } catch (error) {
-        console.error('Error fetching data:', error);
-      }
-    };
+    // const handleSearch = async (query) => {
+    //   console.log(query);
+    //   if (query.trim() === '') return;
+    //   try {
+    //     const response = await fetch(`https://api.lorcana-api.com/cards/fetch?search=Name~${query}`);
+    //     const data = await response.json();
+    //     console.log(data);
+    //     setResults(data);
+    //   } catch (error) {
+    //     console.error('Error fetching data:', error);
+    //   }
+    // };
   
     return (
       <div>
-        <SearchBar onSearch={handleSearch} />
+        {/* <SearchBar onSearch={handleSearch} /> */}
         <div>
           {results.map((result) => (
             <div key={result.id}>{result.name}</div>
